% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz_tteICE.R
\name{basehaz.tteICE}
\alias{basehaz.tteICE}
\title{Baseline hazards of 'tteICE' objects}
\usage{
basehaz.tteICE(object)
}
\arguments{
\item{object}{A fitted object returned by the function \code{tteICE}, \code{surv.tteICE}, or \code{scr.tteICE}.}
}
\value{
A data frame of baseline cumulative hazards in the working Kaplan-Meier or Cox models, stratified by treatment groups.
The first column is time, the following columns are baseline cumulative hazards.
}
\description{
This function extracts the baseline cumulative hazards in the survival models
}
