% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bshaz.R
\name{bshaz}
\alias{bshaz}
\title{S3 method of baseline hazards}
\usage{
bshaz(x)
}
\arguments{
\item{x}{A fitted object returned by the function \code{tteICE}, \code{surv.tteICE}, or \code{scr.tteICE}.}
}
\value{
A data frame of baseline cumulative hazards in the working Kaplan-Meier or Cox models, stratified by treatment groups from the model object.
}
\description{
This function extracts the baseline cumulative hazards in the survival models
}
\keyword{internal}
