% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example data}
\format{
A data frame with variables:
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/9315770}
}
\usage{
dat
}
\description{
{``Simulated data.''}
}
\keyword{datasets}
