% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ICH.R
\name{ICHe9r1Shiny}
\alias{ICHe9r1Shiny}
\title{Shiny for ICHe9r1}
\usage{
ICHe9r1Shiny()
}
\value{
Rshiny inteface
}
\description{
This function opens the Rshiny app for ICHe9r1.
}
\examples{
if(interactive()){
  ICHe9r1Shiny()
}


}
