% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICH_plot_ate.R
\name{plot.ate}
\alias{plot.ate}
\title{plot}
\usage{
\method{plot}{ate}(
  fit,
  decrease = FALSE,
  conf.int = 0.95,
  nboot = 0,
  seed = 0,
  xlab = "Time",
  xlim = NULL,
  ylim = c(-1, 1),
  ...
)
}
\arguments{
\item{fit}{fit}

\item{decrease}{decrease}

\item{conf.int}{conf}

\item{nboot}{nboot}

\item{seed}{seed}

\item{xlab}{xlab}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{...}{Other augments in function \code{\link{plot.default}} or function \code{\link{curve}}}
}
\value{
curves
}
\description{
Plot
}
\examples{
plot(1,1)

}
\seealso{
\code{\link[graphics]{plot.default}},
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
}
