% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ICH.R
\name{plot.ICH}
\alias{plot.ICH}
\title{Plotting the estimated function}
\usage{
\method{plot}{ICH}(
  x,
  type = c("ate", "inc")[1],
  decrease = FALSE,
  conf.int = 0.95,
  nboot = 0,
  seed = 0,
  xlab = "Time",
  xlim = NULL,
  ylim = NULL,
  legend.inc = c("Treated", "Control"),
  col.inc = c("brown", "darkcyan"),
  cex.legend = 0.9,
  ...
)
}
\arguments{
\item{x}{A fitted object from \code{surv.ICH} or \code{scr.ICH}.}

\item{type}{Which plot to create: \code{ate} indicates to plot the estimated treatment effect; \code{inc} indicates to plot the estimated cumulative incidence function.}

\item{decrease}{A logical variable indicating whether displaying the cumulative incidence
function (\code{decrease = FALSE}) or survival function (\code{decrease = TRUE}).}

\item{conf.int}{Level of the confidence interval. If \code{conf.int = NULL}, then the condifence
interval will not be provided.}

\item{nboot}{Number of resamplings in bootstrapping. Default \code{nboot = 0}, using the explicit
formula of the standard error.}

\item{seed}{Seed for bootstrapping.}

\item{xlab}{Label for x-axis.}

\item{xlim}{Limit for x-axis.}

\item{ylim}{Limit for y-axis.}

\item{legend.inc}{Change the legend of the estimated cumulative incidence function plot. Only valid when \code{type=inc}.}

\item{col.inc}{Color of curves in the estimated cumulative incidence function. Only valid when \code{type=inc}.}

\item{cex.legend}{Size of legend. Only valid when \code{type=inc}.}

\item{...}{Other augments in function \code{\link{plot.default}} or function \code{\link{curve}}}
}
\description{
This function plots the estimated potential cumulative incidence function
with pointwise confidence intervals.
}
\examples{
## load data
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
## plot treatment effects with p-values
for (st in c('composite','natural','removed','whileon','principal')){
fit = surv.ICH(A, bmt$t2, bmt$d4, st)
plot(fit, type="inc", ylim=c(0,1))
p = fit$p.val
if (!is.null(p)) text(200, 0.8, paste0('P = ', round(p,3)))
}
## plot counterfactual cumulative incidence functions
for (st in c('composite','natural','removed','whileon','principal')){
fit = surv.ICH(A, bmt$t2, bmt$d4, st)
plot(fit, type="ate", ylim=c(0,1))
}

}
\seealso{
\code{\link{plot_ate}},
\code{\link{plot_inc}}
}
