% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ICH.R
\name{print.ICH}
\alias{print.ICH}
\title{Print the main results of ICHe9r1}
\usage{
\method{print}{ICH}(x, digits = 3, ...)
}
\arguments{
\item{x}{A fitted object from \code{surv.ICH} or \code{scr.ICH}.}

\item{digits}{The digits of results}

\item{...}{Other augments in function \code{\link{print.default}}}
}
\description{
This function summarize the results
}
\examples{
## load data
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
## plot treatment effects with p-values
for (st in c('composite','natural','removed','whileon','principal')){
fit = surv.ICH(A, bmt$t2, bmt$d4, st)
print(fit)
}
## plot counterfactual cumulative incidence functions
for (st in c('composite','natural','removed','whileon','principal')){
fit = surv.ICH(A, bmt$t2, bmt$d4, st)
print(fit, digits=2)
}

}
\seealso{
\code{\link{surv.ICH}},
\code{\link{scr.ICH}}
}
