% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_tteICE.R
\name{print.summary.tteICE}
\alias{print.summary.tteICE}
\title{Print the summary of 'tteICE'}
\usage{
\method{print}{summary.tteICE}(x, digits = 3, ...)
}
\arguments{
\item{x}{A fitted object returned by the function \code{tteICE}, \code{surv.tteICE}, or \code{scr.tteICE}.}

\item{digits}{The digits of the results}

\item{...}{Other arguments in function \code{\link{print.default}}}
}
\value{
Print the summary of a tteICE object
}
\description{
Print the summary of 'tteICE'
}
