% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICH_functions.R
\name{surv.composite}
\alias{surv.composite}
\title{xx}
\usage{
surv.composite(A, Time, cstatus, weights = rep(1, length(A)), subset = NULL)
}
\arguments{
\item{A}{Treatment}

\item{Time}{Time}

\item{cstatus}{Status}

\item{weights}{Weight}

\item{subset}{Subset}
}
\value{
confidence interval,
convergence list,
logit transformed data
}
\description{
xx
}
\details{
\describe{
method
}
}
\examples{
1

}
\seealso{
\code{\link[stats]{uniroot}}
}
