% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_HR.R
\name{surv.HR}
\alias{surv.HR}
\title{Estimating the hazard ratio under ICH E9 (R1)}
\usage{
surv.HR(
  A,
  Time,
  cstatus,
  strategy = "composite",
  cov1 = NULL,
  conf.int = 0.95,
  weights = NULL,
  subset = NULL
)
}
\arguments{
\item{A}{Treatment indicator, 1 for treatment and 0 for control.}

\item{Time}{Time to event.}

\item{cstatus}{Indicator of event, 1 for the primary event, 2 for the intercurrent event, 0 for censoring.}

\item{strategy}{Strategy to address intercurrent events, \code{"treatment"} indicating treatment policy strategy,
\code{"composite"} indicating composite variable strategy, \code{"natural"} indicating hypothetical strategy
(Scenario I, controlling the hazard of intercurrent events), \code{"removed"} indicating hypothetical strategy
(Scenario II, removing intercurrent events), and \code{"whileon"} indicating while on treatment strategy.}

\item{cov1}{Baseline covariates.}

\item{weights}{Weight for each subject (not applied to the while on treatment strategy).}

\item{subset}{Subset, either numerical or logical.}

\item{cnf.int}{Level of the confidence interval.}
}
\value{
A list including
\describe{
\item {logHR} {Estimated log hazard ratio (logHR) of the treatment effect on the primary event.}
\item {se} {Standard error of the estimated log hazard ratio (logHR).}
\item {CI} {Confidence interval of the hazard ratio (HR).}
\item {p.val} {P value of the hazard ratio.}
}
}
\description{
This function estimates the hazard ratio
for time-to event data under ICH E9 (R1) to address intercurrent events. Multiple
strategies except the principal stratum strategy are allowed.
}
\details{
\describe{
For the treatment policy and hypothetical strategies, the hazard ratio (HR) is given by the Cox
regression regarding intercurrent events as censoring. For the composite variable strategy, the
hazard ratio is given by the the Cox regression regarding the first occurrence of either intercurrent
event or primary event as the event of interest. For the while on treatment strategy, the hazard
ratio is given by the Fine-Gray subdistribution model. There is no existing method to estimate the
hazard ratio using principal stratum strategy. \n
The weakness of using hazard ratio to infer treatment effects is critical. First, the hazard ratio
relies on model specification. Second, the hazard ratio is not collapsible. Therefore, the hazard
ratio should only be treated as a descriptive or exploratory measure of the treatment effect.
}
}
\examples{
\code{
## Generate simulated data
dat = generatedata(500)
## composite variable strategy
fit = surv.HR(dat$Z, dat$Time, dat$cstatus, "composite")
## while on treatment strategy
fit = surv.HR(dat$Z, dat$Time, dat$cstatus, "whileon", cov1=dat$X)
}

}
\seealso{
\code{\link[ICHe9r1]{surv.ICH}}
}
