% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ice.R
\name{Surv.ice}
\alias{Surv.ice}
\title{Outcome def}
\usage{
Surv.ice(time.p, status.p, time.i = NULL, status.i = NULL)
}
\arguments{
\item{time.p}{Treatment indicator, 1 for treatment and 0 for control.}

\item{status.p}{Time to the primary (terminal) event.}

\item{time.i}{Indicator of the primary (terminal) event, 1 for event and 0 for censoring.}

\item{status.i}{Time to the intercurrent event.}
}
\value{
A list including the fitted object and input variables.
}
\description{
TOutcome def
}
\examples{
data(bmt)
bmt = transform(bmt, d4=d2+d3)
with(bmt,Surv.ice(t1,d1))
with(bmt,Surv.ice(t1,d4))
with(bmt,Surv.ice(t1,d1,t2,d2))
}
