% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICH_functions.R
\name{surv.whileon.eff}
\alias{surv.whileon.eff}
\title{xx}
\usage{
surv.whileon.eff(A, Time, cstatus, X = NULL, subset = NULL)
}
\arguments{
\item{A}{Treatment}

\item{Time}{Time}

\item{cstatus}{Status}

\item{subset}{Subset}

\item{weights}{Weight}
}
\value{
confidence interval,
convergence list,
logit transformed data
}
\description{
xx
}
\details{
\describe{
method
}
}
\examples{
1

}
\seealso{
\code{\link[stats]{uniroot}}
}
