% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tteICE-package.R
\docType{package}
\name{tteICE-package}
\alias{tteICE-package}
\title{tteICE: Treatment Effect Estimation for Time-to-Event Data with Intercurrent Events}
\description{
Main functions:
\itemize{
\item \code{\link[tteICE]{tteICE}} Fit the CIF for time-to-event data using formula
\item \code{\link[tteICE]{scr.tteICE}} Fit the CIF for time-to-event data with intercurrent events for semicompeting risks data
\item \code{\link[tteICE]{surv.tteICE}} Fit the CIF for time-to-event with intercurrent events for competing risks data
\item \code{\link[tteICE]{plot.tteICE}} Graphical results of tteICE
\item \code{\link[tteICE]{print.tteICE}} Print a short summary of the estimated treatment effect
\item \code{\link[tteICE]{summary.tteICE}} Summary of tteICE
\item \code{\link[tteICE]{predict.tteICE}} Risk prediction at specific time points
\item \code{\link[tteICE]{tteICEShiny}} Interactive Shiny app for tteICE
}
Example data:
\itemize{
\item \code{\link{bmt}} Data from Section 1.3 of Klein and Moeschberger (1997)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mephas/tteICE}
  \item \url{https://mephas.github.io/tteICE/}
  \item Report bugs at \url{https://github.com/mephas/tteICE/issues}
}

}
\author{
\strong{Maintainer}: Yi Zhou \email{yzhou@pku.edu.cn}

Authors:
\itemize{
  \item Yuhao Deng \email{dengyuhao@pku.edu.cn}
}

}
