% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ICH.R
\name{tteICEShiny}
\alias{tteICEShiny}
\title{Shiny for tteICE}
\usage{
ICHe9r1Shiny()
}
\value{
Rshiny inteface
}
\description{
This function opens the Rshiny app for ICHe9r1.
}
\examples{
if(interactive()){
  ICHe9r1Shiny()
}


}
