% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ate.R
\name{plot.ate}
\alias{plot.ate}
\title{Plotting the estimated treatment effect}
\usage{
\method{plot}{ate}(
  fit,
  decrease = FALSE,
  conf.int = 0.95,
  nboot = 0,
  seed = 0,
  xlab = "Time",
  xlim = NULL,
  ylim = c(-1, 1),
  ...
)
}
\arguments{
\item{fit}{A fitted object from \code{surv.ICH}.}

\item{decrease}{A logical variable indicating whether displaying the difference in cumulative
incidence functions (\code{decrease = FALSE}) or survival functions (\code{decrease = TRUE}).}

\item{conf.int}{Level of the confidence interval. If \code{conf.int = NULL}, then the condifence
interval will not be provided.}

\item{nboot}{Number of resamplings in bootstrapping. Default \code{nboot = 0}, using the explicit
formula of the standard error.}

\item{seed}{Seed for bootstrapping.}

\item{...}{Other augments in function \code{\link{plot.default}} or function \code{\link{curve}}}
}
\description{
This function plots the estimated treatment effect (difference in potential
cumulative incidences under treated and control) with pointwise confidence intervals.
}
\examples{
\code{
dat = generatedata(500)
for (st in c('composite','natural','removed','whileon','principal')){
fit = surv.ICH(dat$Z, dat$Time, dat$cstatus, st)
plot.ate(fit, ylim=c(-0.4,0.4))
}
}

}
\seealso{
\code{\link[graphics]{plot.default}},
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
}
