% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ICH.R
\name{plot.ICH}
\alias{plot.ICH}
\title{Plotting the estimated function}
\usage{
\method{plot}{ICH}(
  x,
  type = c("ate", "inc")[1],
  decrease = FALSE,
  conf.int = 0.95,
  nboot = 0,
  seed = 0,
  xlab = "Time",
  xlim = NULL,
  ylim = NULL,
  legend = c("Treated", "Control"),
  cex = 0.9,
  ...
)
}
\arguments{
\item{x}{A fitted object from \code{surv.ICH}.}

\item{type}{which plot to create}

\item{decrease}{A logical variable indicating whether displaying the cumulative incidence
function (\code{decrease = FALSE}) or survival function (\code{decrease = TRUE}).}

\item{conf.int}{Level of the confidence interval. If \code{conf.int = NULL}, then the condifence
interval will not be provided.}

\item{nboot}{Number of resamplings in bootstrapping. Default \code{nboot = 0}, using the explicit
formula of the standard error.}

\item{seed}{Seed for bootstrapping.}

\item{xlab}{Label for x-axis.}

\item{xlim}{Limit for x-axis.}

\item{ylim}{Limit for y-axis.}

\item{legend}{Legend.}

\item{cex}{Size of legend.}

\item{...}{Other augments in function \code{\link{plot.default}} or function \code{\link{curve}}}
}
\description{
This function plots the estimated potential cumulative incidence function
with pointwise confidence intervals.
}
\examples{
## dat = .generatedata(500)
for (st in c('composite')){
fit = surv.ICH(dat$Z, dat$Time, dat$cstatus, st)
plot(fit, type="ate", ylim=c(0,1))
}

for (st in c('composite')){
fit = surv.ICH(dat$Z, dat$Time, dat$cstatus, st)
plot(fit, type="inc", ylim=c(0,1))
p = fit$p.val
if (!is.null(p)) text(0.6, 0.8, paste0('P = ', round(p,3)))
}

}
\seealso{
\code{\link{plot_ate}},
\code{\link{plot_inc}}
}
