% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inc.R
\name{plot_inc}
\alias{plot_inc}
\title{Plotting the estimated cumulative incidence function}
\usage{
plot_inc(
  fit,
  decrease = FALSE,
  conf.int = 0.95,
  nboot = 0,
  seed = 0,
  xlab = "Time",
  xlim = NULL,
  ylim = c(0, 1),
  legend = c("Treated", "Control"),
  cex = 0.9,
  ...
)
}
\arguments{
\item{fit}{A fitted object from \code{surv.ICH}.}

\item{decrease}{A logical variable indicating whether to display the cumulative incidence
function (\code{decrease = FALSE}) or survival function (\code{decrease = TRUE}).}

\item{conf.int}{Level of the confidence interval. If \code{conf.int = NULL}, then the condifence
interval will not be provided.}

\item{nboot}{Number of resamplings in bootstrapping. Default \code{nboot = 0}, using the explicit
formula of the standard error.}

\item{seed}{Seed for bootstrapping.}

\item{xlab}{Label for x-axis.}

\item{xlim}{Limit for x-axis.}

\item{ylim}{Limit for y-axis.}

\item{legend}{Legend.}

\item{cex}{Size of legend.}

\item{...}{Other augments in function \code{\link{plot.default}} or function \code{\link{curve}}}
}
\description{
This function plots the estimated potential cumulative incidence function
with pointwise confidence intervals.
}
\examples{
## load data and fit the model
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
fit = surv.ICH(A, bmt$t2, bmt$d4, 'composite')
## plot asymptotic confidence intervals based on explicit formulas
plot_inc(fit, legend=c('AML','ALL'), ylim=c(0,1))
## plot bootstrap confidence intervals
plot_inc(fit, nboot=200, legend=c('AML','ALL'), ylim=c(0,1))

}
\seealso{
\code{\link[graphics]{plot.default}},
\code{\link[graphics]{points}},
\code{\link[graphics]{curve}},
}
