% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tteICE.R
\name{predict.tteICE}
\alias{predict.tteICE}
\title{Risk prediction at specific time points}
\usage{
\method{predict}{tteICE}(object, timeset = NULL, ...)
}
\arguments{
\item{object}{A fitted object returned by the function \code{surv.tteICE} or \code{scr.tteICE}.}

\item{timeset}{Time at which to predict the risk. If \code{timeset=NULL}, risks will be predict at the
quartiles of the maximum follow-up time.}

\item{...}{Other augments in function \code{\link[stats]{predict}}}
}
\value{
A matrix. The meanings of each row are: time points, potential cumulative incidences (under
treated and under control), treatment effects, standard errors, and P-values.

predict a tteICE object.
The meanings of each row are: time points, potential cumulative incidences (under
treated and under control), treatment effects, standard errors, and P-values.
}
\description{
This function predicts the potential cumulative incidence function and treatment effect at
specific time points.
}
\examples{
## load data
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
X = as.matrix(bmt[,c('z1','z3','z5')])
## Composite variable strategy,
## nonparametric estimation without covariates
fit1 = scr.tteICE(A, bmt$t1, bmt$d1, bmt$t2, bmt$d2, "composite")
predict(fit1, timeset=c(670,2000))

fit2 = tteICE(Surv.ice(t2, d4)~A, data=bmt,
strategy="composite", cov.formula=~z1+z3+z5, method='eff')
predict(fit2, timeset=c(670,2000))
predict(fit2)


}
\seealso{
\code{\link[tteICE]{scr.tteICE}}, \code{\link[tteICE]{surv.tteICE}}, \code{\link[tteICE]{surv.boot}}
}
