% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskpredict.R
\name{riskpredict}
\alias{riskpredict}
\title{Risk prediction at specific time points}
\usage{
riskpredict(fit, timeset = NULL, nboot = 0, seed = 0)
}
\arguments{
\item{fit}{An ICH object.}

\item{timeset}{Time at which to predict the risk. If \code{timeset=NULL}, risks will be predict at the 
quartiles of the maximum follow-up time.}

\item{nboot}{Number of resamplings in bootstrapping. Default \code{nboot = 0}, using the explicit
formula of the standard error.}

\item{seed}{Seed for bootstrapping.}
}
\value{
A matrix. The meanings of each row are: time points, potential cumulative incidences (under 
treated and under control), treatment effects, standard errors, P-values.
}
\description{
This function predicts the potential cumulative incidence function and treatment effect at
specific time points.
}
\seealso{
\code{\link[ICHe9r1]{scr.ICH}}, \code{\link[ICHe9r1]{surv.ICH}}, \code{\link[ICHe9r1]{surv.boot}}
}
