% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tteICE.R
\name{summary.tteICE}
\alias{summary.tteICE}
\title{Summary of tteICE}
\usage{
\method{summary}{tteICE}(object, ...)
}
\arguments{
\item{object}{A fitted object returned by the function \code{surv.tteICE} or \code{scr.tteICE}.}

\item{...}{Other augments in function \code{\link{print.default}}}
}
\value{
A list that consists of summaries of a tteICE object: data type, strategy, estimation method, maximum follow-up time,
sample size, treated sample size, controlled sample size, p-value, and predicted risks at quartiles
}
\description{
This function summarizes the results
}
\examples{
## load data
data(bmt)
bmt = transform(bmt, d4=d2+d3)
A = as.numeric(bmt$group>1)
## summarize the results
fit = surv.tteICE(A, bmt$t2, bmt$d4, 'natural')
summary(fit)

}
\seealso{
\code{\link[tteICE]{surv.tteICE}},
\code{\link[tteICE]{scr.tteICE}},
\code{\link[tteICE]{print.tteICE}}
}
