% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tteICEShiny.R
\name{tteICEShiny}
\alias{tteICEShiny}
\title{Shiny app for tteICE}
\usage{
tteICEShiny()
}
\value{
Rshiny inteface
}
\description{
This function opens the Rshiny app for tteICE. R Shiny application can be used for generating plots and
basic analysis results. It provides a point-and-click interface, so users can
obtain results without writing R code directly.
}
\examples{
if(interactive()){
  tteICEShiny()
}


}
